Move board logic out of activity
