+++ app/src/main/java/com/tavarus/artabletop/Controllers/AuthController.kt
@@ -48,4 +48,8 @@ class AuthController() {
         activity.onLogout()
     }
 
+    fun getUser() : FirebaseUser? {
+        return auth.currentUser
+    }
+
 }
